Distributed training revealed some parameters are not tracked during backwards propagation in single-gpu training. Also added class weights for Mapillary dataset to the conditioned head for balancing dataset class imbalance.
Since distributed training detects this oversight, needed to add a new auxiliary module called SBDModule, a variant of DModule. The only difference is that the last bottleneck layer of d_branch layers are not initialized or used in the forward method, since conditioned models use only temp_d and not x_d. Conversely, x_d is used in fusion-based heads like Bag, so it does require the last bottleneck layer of d_branch layers.